refactor(topology): move gossip serializer to new package for reusing
