+++ topology/src/main/java/io/camunda/zeebe/topology/gossip/ClusterTopologyGossiper.java
@@ -12,6 +12,7 @@ import io.atomix.cluster.Member;
 import io.atomix.cluster.MemberId;
 import io.atomix.cluster.messaging.ClusterCommunicationService;
 import io.camunda.zeebe.scheduler.ConcurrencyControl;
+import io.camunda.zeebe.topology.serializer.ClusterTopologyGossipSerializer;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.topology.state.ClusterTopology;
 import java.util.Collections;
