+++ topology/src/main/java/io/camunda/zeebe/topology/serializer/ClusterTopologyGossipSerializer.java
@@ -5,18 +5,13 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.topology.gossip;
+package io.camunda.zeebe.topology.serializer;
 
-interface ClusterTopologyGossipSerializer {
+import io.camunda.zeebe.topology.gossip.ClusterTopologyGossipState;
+
+public interface ClusterTopologyGossipSerializer {
 
   byte[] encode(ClusterTopologyGossipState gossipState);
 
   ClusterTopologyGossipState decode(byte[] encodedState);
-
-  final class DecodingFailed extends RuntimeException {
-
-    public DecodingFailed(final Throwable cause) {
-      super(cause);
-    }
-  }
 }
