+++ topology/src/main/java/io/camunda/zeebe/topology/serializer/ProtoBufSerializer.java
@@ -5,10 +5,11 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.topology.gossip;
+package io.camunda.zeebe.topology.serializer;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 import io.atomix.cluster.MemberId;
+import io.camunda.zeebe.topology.gossip.ClusterTopologyGossipState;
 import io.camunda.zeebe.topology.protocol.Topology;
 import io.camunda.zeebe.topology.state.ClusterChangePlan;
 import io.camunda.zeebe.topology.state.ClusterTopology;
@@ -52,17 +53,6 @@ public class ProtoBufSerializer implements ClusterTopologyGossipSerializer {
     return clusterTopologyGossipState;
   }
 
-  private Topology.ClusterTopology encodeClusterTopology(
-      final io.camunda.zeebe.topology.state.ClusterTopology clusterTopology) {
-    final var members =
-        clusterTopology.members().entrySet().stream()
-            .collect(Collectors.toMap(e -> e.getKey().id(), e -> encodeMemberState(e.getValue())));
-    return Topology.ClusterTopology.newBuilder()
-        .setVersion(clusterTopology.version())
-        .putAllMembers(members)
-        .build();
-  }
-
   private io.camunda.zeebe.topology.state.ClusterTopology decodeClusterTopology(
       final Topology.ClusterTopology encodedClusterTopology) {
 
@@ -72,6 +62,17 @@ public class ProtoBufSerializer implements ClusterTopologyGossipSerializer {
             .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
     return new io.camunda.zeebe.topology.state.ClusterTopology(
         encodedClusterTopology.getVersion(), members, ClusterChangePlan.empty());
+  }
+
+  private Topology.ClusterTopology encodeClusterTopology(
+      final io.camunda.zeebe.topology.state.ClusterTopology clusterTopology) {
+    final var members =
+        clusterTopology.members().entrySet().stream()
+            .collect(Collectors.toMap(e -> e.getKey().id(), e -> encodeMemberState(e.getValue())));
+    return Topology.ClusterTopology.newBuilder()
+        .setVersion(clusterTopology.version())
+        .putAllMembers(members)
+        .build();
   }
 
   private io.camunda.zeebe.topology.state.MemberState decodeMemberState(
